+++ client/src/components/Publications/Publicar/publicar.css
@@ -5,6 +5,7 @@
   width: 100%;
   height: 100%;
   background-color: rgba(0, 0, 0, 0.5);
+  z-index: 999;
   /* Fondo semi-transparente */
   display: flex;
   justify-content: center;
