+++ client/src/components/Publications/Publications.css
@@ -171,7 +171,6 @@
 
 .love-button,
 .comment-button {
-  position: relative;
   background-color: transparent;
   border: none;
   cursor: pointer;
