+++ client/src/components/TopBar/TopBar.css
@@ -8,15 +8,58 @@
   position: fixed;
   box-shadow: 5px 3px 5px black;
 }
+.top-bar-start{
+  gap: 30%;
+}
 
 .img-top-bar {
   margin-left: 1.75rem;
   width: 100px;
-  height: 22px;
+  height: 24px;
 }
 
 @media (min-width: 768px) {
   .hide-sidebar-button {
     display: none;
   }
+}
+
+.searchAndImg {
+  width: 40%;
+  display: flex;
+  flex-direction: row;
+}
+
+.linkImg {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.header-content {
+  width: 50%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.header-content i {
+  color: #fff;
+  font-size: 20px;
+  margin-left: 1rem;
+}
+
+.header-content .btn {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 10px;
+  border-radius: 8px;
+  background: #242424;
+  margin-left: 1rem;
+}
+
+.header-content .btn i {
+  color: hsla(0, 0%, 100%, 1.6);
+  margin-right: 10px;
 }
